refactor(frontend): rename helpers to api and simplify auth route wiring
