+++ src/pages/HomeAdmin.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
 import TableFila from "../components/TableFila";
 import CardFilaNow from "../components/CardFilaNow";
-import { getFila } from "../helpers/filaApi"; // Importamos los datos de fila
-import { atenderTurnoConId, putFinalizarAtencion } from "../helpers/filaApi"; // Importamos la función para atender un turno
-import { postTurnoEnFila } from "../helpers/filaApi"; // Importamos la función para agregar un turno a la fila
+import { getFila } from "../api/filaApi"; // Importamos los datos de fila
+import { atenderTurnoConId, putFinalizarAtencion } from "../api/filaApi"; // Importamos la función para atender un turno
+import { postTurnoEnFila } from "../api/filaApi"; // Importamos la función para agregar un turno a la fila
 import { useNavigate } from "react-router-dom";
-import { verificarToken } from "../helpers/login"; // Importamos la función para verificar el token
+import { verificarToken } from "../api/login"; // Importamos la función para verificar el token
 import Swal from "sweetalert2";
