+++ src/pages/LoginAdmin.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import Swal from "sweetalert2";
 
 import "../css/login.css";
-import { LogIn } from "../helpers/login"; // Import the getToken function
+import { LogIn } from "../api/login"; // Import the getToken function
 
 const LoginAdmin = ({ cambiarLogin }) => {
   const navigate = useNavigate();
