+++ src/pages/TurnosChart.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Bar } from "react-chartjs-2";
 import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend } from "chart.js";
 import { FaFilter } from "react-icons/fa";
-import {getTurnos} from "../helpers/turnosApi";
+import {getTurnos} from "../api/turnosApi";
 ChartJS.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend);
